migrate from nuxt-cron to native nitro tasks
