+++ package.json
@@ -58,7 +58,6 @@
     "drizzle-kit": "^0.30.5",
     "esbuild": "^0.24.0",
     "happy-dom": "^17.4.4",
-    "nuxt-cron": "^1.8.0",
     "prettier": "^3.5.3",
     "tailwindcss": "^4.0.9",
     "tsx": "^4.19.3",
