--- server/cron/clean-flairs.ts
@@ -1,35 +1,0 @@
-import { defineCronHandler } from '#nuxt/cron';
-import {
-  itemsTable,
-  purchasesTable,
-  userSettingsTable,
-} from '~/database/schema';
-import { db } from '../db';
-import { and, sql, eq, notInArray, gte, isNotNull } from 'drizzle-orm';
-
-export default defineCronHandler('everyFiveMinutes', async () => {
-  await db
-    .update(userSettingsTable)
-    .set({ flair_id: null })
-    .where(
-      and(
-        isNotNull(userSettingsTable.flair_id),
-        notInArray(
-          userSettingsTable.user_id,
-          db
-            .select({
-              user_id: purchasesTable.user_id,
-            })
-            .from(purchasesTable)
-            .innerJoin(itemsTable, eq(purchasesTable.item_id, itemsTable.id))
-            .where(
-              and(
-                eq(userSettingsTable.user_id, purchasesTable.user_id),
-                gte(purchasesTable.purchased_on, sql`NOW() - INTERVAL 7 DAY`),
-                eq(itemsTable.category, 'flair')
-              )
-            )
-        )
-      )
-    );
-});
