--- server/cron/clean-hatchery.ts
@@ -1,6 +1,0 @@
-import { defineCronHandler } from '#nuxt/cron';
-import { cleanUp } from '~/server/clean-up';
-
-export default defineCronHandler('everyFiveMinutes', () => {
-  cleanUp();
-});
